Refactor: PostHeader컴포넌트에서 style분리
